Рефакторинг
- Переделан инжектинг всех зависимых сервисов
- Извлечен компьютер контента строковых литералов
