+++ bundles/com.e1c.ssl.bsl/src/com/e1c/ssl/bsl/CommonFunctionFixedDataTypesComputer.java
@@ -27,6 +27,8 @@ import org.eclipse.xtext.util.Pair;
 import com._1c.g5.v8.dt.bsl.model.Expression;
 import com._1c.g5.v8.dt.bsl.model.ExtendedCollectionType;
 import com._1c.g5.v8.dt.bsl.model.Invocation;
+import com._1c.g5.v8.dt.bsl.resource.DynamicFeatureAccessComputer;
+import com._1c.g5.v8.dt.bsl.resource.TypesComputer;
 import com._1c.g5.v8.dt.mcore.DerivedProperty;
 import com._1c.g5.v8.dt.mcore.Environmental;
 import com._1c.g5.v8.dt.mcore.McoreFactory;
@@ -38,7 +40,9 @@ import com._1c.g5.v8.dt.mcore.TypeItem;
 import com._1c.g5.v8.dt.mcore.util.McoreUtil;
 import com._1c.g5.v8.dt.platform.IEObjectProvider;
 import com._1c.g5.v8.dt.platform.IEObjectTypeNames;
+import com._1c.g5.v8.dt.platform.version.IRuntimeVersionSupport;
 import com._1c.g5.v8.dt.platform.version.Version;
+import com.google.inject.Inject;
 
 /**
  * Extension computer of invocation types of 1C:SSL API module function {@code Common.FixedData()} that
@@ -50,6 +54,15 @@ import com._1c.g5.v8.dt.platform.version.Version;
 public class CommonFunctionFixedDataTypesComputer
     extends AbstractCommonModuleObjectAttributeValueTypesComputer
 {
+    private final TypesComputer typesComputer;
+
+    @Inject
+    public CommonFunctionFixedDataTypesComputer(TypesComputer typesComputer, IRuntimeVersionSupport versionSupport,
+        DynamicFeatureAccessComputer dynamicFeatureAccessComputer)
+    {
+        super(typesComputer, versionSupport, dynamicFeatureAccessComputer);
+        this.typesComputer = typesComputer;
+    }
 
     @Override
     public List<TypeItem> getTypes(Invocation inv)
@@ -66,7 +79,7 @@ public class CommonFunctionFixedDataTypesComputer
 
         Environmental envs = EcoreUtil2.getContainerOfType(expr, Environmental.class);
 
-        List<TypeItem> types = this.getTypesComputer().computeTypes(expr, envs.environments());
+        List<TypeItem> types = typesComputer.computeTypes(expr, envs.environments());
         if (types.isEmpty())
             return Collections.emptyList();
