+++ bundles/com.e1c.ssl.bsl/src/com/e1c/ssl/bsl/CommonFunctionUnloadColumnTypesComputer.java
@@ -22,6 +22,9 @@ import org.eclipse.xtext.util.Pair;
 import com._1c.g5.v8.dt.bsl.model.Expression;
 import com._1c.g5.v8.dt.bsl.model.ExtendedCollectionType;
 import com._1c.g5.v8.dt.bsl.model.Invocation;
+import com._1c.g5.v8.dt.bsl.model.StringLiteral;
+import com._1c.g5.v8.dt.bsl.resource.DynamicFeatureAccessComputer;
+import com._1c.g5.v8.dt.bsl.resource.TypesComputer;
 import com._1c.g5.v8.dt.bsl.typesystem.util.TypeSystemUtil;
 import com._1c.g5.v8.dt.mcore.Environmental;
 import com._1c.g5.v8.dt.mcore.McorePackage;
@@ -29,8 +32,10 @@ import com._1c.g5.v8.dt.mcore.Property;
 import com._1c.g5.v8.dt.mcore.Type;
 import com._1c.g5.v8.dt.mcore.TypeItem;
 import com._1c.g5.v8.dt.platform.IEObjectProvider;
+import com._1c.g5.v8.dt.platform.version.IRuntimeVersionSupport;
 import com._1c.g5.v8.dt.platform.version.Version;
 import com.google.common.collect.Lists;
+import com.google.inject.Inject;
 
 /**
  * Extension computer of invocation types of 1C:SSL API module function {@code Common.UnloadColumn()} that
@@ -42,6 +47,19 @@ import com.google.common.collect.Lists;
 public class CommonFunctionUnloadColumnTypesComputer
     extends AbstractCommonModuleObjectAttributeValueTypesComputer
 {
+
+    private final TypesComputer typesComputer;
+
+    private final ExpressionValueComputer expressionValueComputer;
+
+    @Inject
+    public CommonFunctionUnloadColumnTypesComputer(TypesComputer typesComputer, IRuntimeVersionSupport versionSupport,
+        DynamicFeatureAccessComputer dynamicFeatureAccessComputer, ExpressionValueComputer expressionValueComputer)
+    {
+        super(typesComputer, versionSupport, dynamicFeatureAccessComputer);
+        this.typesComputer = typesComputer;
+        this.expressionValueComputer = expressionValueComputer;
+    }
 
     @Override
     public List<TypeItem> getTypes(Invocation inv)
@@ -61,7 +79,7 @@ public class CommonFunctionUnloadColumnTypesComputer
             return Collections.emptyList();
 
         Environmental envs = EcoreUtil2.getContainerOfType(dataExpression, Environmental.class);
-        List<TypeItem> types = this.getTypesComputer().computeTypes(dataExpression, envs.environments());
+        List<TypeItem> types = typesComputer.computeTypes(dataExpression, envs.environments());
         if (types.isEmpty())
             return Collections.emptyList();
 
@@ -82,8 +100,9 @@ public class CommonFunctionUnloadColumnTypesComputer
             Collection<Property> properties = all.getFirst();
             for (Property property : properties)
             {
-                String expressionContent = getExpressionContent(nameExpression);
-                if (property.getName().equals(expressionContent))
+                Pair<String, Collection<StringLiteral>> expressionContent =
+                    expressionValueComputer.getExpressionContent(nameExpression);
+                if (expressionContent != null && property.getName().equals(expressionContent.getFirst()))
                 {
                     IEObjectProvider provider = IEObjectProvider.Registry.INSTANCE.get(McorePackage.Literals.TYPE_ITEM,
                         versionSupport.getRuntimeVersionOrDefault(inv, Version.LATEST));
