+++ bundles/com.e1c.ssl.bsl/src/com/e1c/ssl/bsl/CommonFunctionValueTableToArrayTypesComputer.java
@@ -24,6 +24,8 @@ import com._1c.g5.v8.dt.bsl.model.Expression;
 import com._1c.g5.v8.dt.bsl.model.ExtendedCollectionType;
 import com._1c.g5.v8.dt.bsl.model.Invocation;
 import com._1c.g5.v8.dt.bsl.model.StaticFeatureAccess;
+import com._1c.g5.v8.dt.bsl.resource.DynamicFeatureAccessComputer;
+import com._1c.g5.v8.dt.bsl.resource.TypesComputer;
 import com._1c.g5.v8.dt.bsl.typesystem.util.TypeSystemUtil;
 import com._1c.g5.v8.dt.mcore.DerivedProperty;
 import com._1c.g5.v8.dt.mcore.Environmental;
@@ -35,8 +37,10 @@ import com._1c.g5.v8.dt.mcore.TypeContainerRef;
 import com._1c.g5.v8.dt.mcore.TypeItem;
 import com._1c.g5.v8.dt.platform.IEObjectProvider;
 import com._1c.g5.v8.dt.platform.IEObjectTypeNames;
+import com._1c.g5.v8.dt.platform.version.IRuntimeVersionSupport;
 import com._1c.g5.v8.dt.platform.version.Version;
 import com.google.common.collect.Lists;
+import com.google.inject.Inject;
 
 /**
  *
@@ -49,6 +53,16 @@ import com.google.common.collect.Lists;
 public class CommonFunctionValueTableToArrayTypesComputer
     extends AbstractCommonModuleObjectAttributeValueTypesComputer
 {
+
+    private final TypesComputer typesComputer;
+
+    @Inject
+    public CommonFunctionValueTableToArrayTypesComputer(TypesComputer typesComputer,
+        IRuntimeVersionSupport versionSupport, DynamicFeatureAccessComputer dynamicFeatureAccessComputer)
+    {
+        super(typesComputer, versionSupport, dynamicFeatureAccessComputer);
+        this.typesComputer = typesComputer;
+    }
 
     @Override
     public List<TypeItem> getTypes(Invocation inv)
@@ -64,7 +78,7 @@ public class CommonFunctionValueTableToArrayTypesComputer
         {
             Environmental envs = EcoreUtil2.getContainerOfType(expr, Environmental.class);
 
-            List<TypeItem> types = this.getTypesComputer().computeTypes(expr, envs.environments());
+            List<TypeItem> types = typesComputer.computeTypes(expr, envs.environments());
             if (types.isEmpty())
                 return Collections.emptyList();
