+++ bundles/com.e1c.ssl.bsl/src/com/e1c/ssl/bsl/internal/ExternalDependenciesModule.java
@@ -13,7 +13,13 @@
 package com.e1c.ssl.bsl.internal;
 
 import org.eclipse.core.runtime.Plugin;
+import org.eclipse.emf.common.util.URI;
+import org.eclipse.xtext.resource.IResourceServiceProvider;
+import org.eclipse.xtext.scoping.IScopeProvider;
 
+import com._1c.g5.v8.dt.bsl.resource.DynamicFeatureAccessComputer;
+import com._1c.g5.v8.dt.bsl.resource.TypesComputer;
+import com._1c.g5.v8.dt.bsl.typesystem.ValueTableDynamicContextDefProvider;
 import com._1c.g5.v8.dt.core.platform.IBmModelManager;
 import com._1c.g5.v8.dt.core.platform.IResourceLookup;
 import com._1c.g5.v8.dt.platform.version.IRuntimeVersionSupport;
@@ -41,6 +47,15 @@ class ExternalDependenciesModule
         bind(IResourceLookup.class).toService();
         bind(IRuntimeVersionSupport.class).toService();
         bind(IBmModelManager.class).toService();
+
+        URI uri = URI.createURI("*.bsl"); //$NON-NLS-1$
+        final IResourceServiceProvider rsp = IResourceServiceProvider.Registry.INSTANCE.getResourceServiceProvider(uri);
+
+        bind(TypesComputer.class).toProvider(() -> rsp.get(TypesComputer.class));
+        bind(DynamicFeatureAccessComputer.class).toProvider(() -> rsp.get(DynamicFeatureAccessComputer.class));
+        bind(IScopeProvider.class).toProvider(() -> rsp.get(IScopeProvider.class));
+        bind(ValueTableDynamicContextDefProvider.class)
+            .toProvider(() -> rsp.get(ValueTableDynamicContextDefProvider.class));
 
     }
